+++ source/timelinelib/wxgui/dialogs/preferences/controller.py
@@ -92,6 +92,12 @@ class PreferencesDialogController(Controller):
             self.config.event_font = font.serialize()
             self.view.SetEventFont(font)
 
+    def on_era_click(self, evt):
+        font = deserialize_font(self.config.era_font)
+        if self.view.ShowEditFontDialog(font):
+            self.config.era_font = font.serialize()
+            self.view.SetEraFont(font)
+
     def on_major_strip_click(self, event):
         font = deserialize_font(self.config.major_strip_font)
         if self.view.ShowEditFontDialog(font):
@@ -186,6 +192,7 @@ class PreferencesDialogController(Controller):
         self.view.SetLegendFont(deserialize_font(self.config.legend_font))
         self.view.SetBalloonFont(deserialize_font(self.config.balloon_font))
         self.view.SetEventFont(deserialize_font(self.config.event_font))
+        self.view.SetEraFont(deserialize_font(self.config.era_font))
         self.view.SetLegendPos(self.config.legend_pos)
         self.view.SetTimeScalePos(self.config.time_scale_pos)
         self.view.SetWorkdayLength(self.config.workday_length)